+++ contracts/interfaces/IApplication.sol
@@ -5,5 +5,5 @@ interface IApplication {
 }
 
 interface IRegisterApplication {
-    function registerApplication(uint256 protocolId, address protocolAddress) external;
+    function registerApplication(uint16 protocolId, address protocolAddress) external;
 }
